Property to disable percentile calculations https://github.com/Netflix/Hystrix/issues/15
The property actually already existed from earlier work I did when opensourcing and was already preventing 'addValue' from being called but I've changed it now to be part of the HystrixRollingPercentile itself.

This makes it so anything that calls it, including Metrics polling won't trigger the internal logic and when disabled it will just do no-op and return -1 for values.