Add OnCancel method to handle cleanup of blocking libs
This change extends the lifecycles of Hystrix commands (both imperative
and reactive styles) to execute the AbstractCommand.onCancel() method
when a timeout occurs or the method Future.cancel(boolean) is invoked
on a Future obtained via HystrixCommand.queue().

The aim of this extension of the lifecycle is to allow easy cleanup of
blocking calls on execution threads that do not react to interruption,
e.g., java.net.Socket.read() on some operative systems that do not
support IO interruption and java.sql.Statement.execute().

The typical logic to be run in an onCancel method is failure prone,
likely to throw exceptions like IOException. In case of a throwable
propagating outside of an onCancel invocation, the failure will be
propagated gto the outside as FailureType.COMMAND_EXCEPTION, with the
cause of the HystrixRuntimeException being a HystrixOnCancelException,
which ultimately wraps the throwable propagated outside of the
onCancel invocation.

In case of a timeout occurring, the execution of AbstractCommand.onCancel()
occurs on the Hystrix timer thread.

In case of HystrixCommand.queue().cancel(true), AbstractCommand.onCancel()
is invoked from the thread invoking Future.cancel(boolean).

To be considered: Add a failure type for OnCancel failure instead of wrapping
      HystrixOnCancelException ?
Follow-ups: Add support for hooks

Closes #1643