[HYSTRIX-1721] - Fixes normal-case unsubscribes
Operations on HysterixObservableCommand treat all unsubscriptions as unsuccessful. This should not be the case when an unsubscription is due to normal circumstances, e.g. zipping an observable that emits more items than the other one (see added test case in HysterixCircuitBreakerTest.

This patch marks a HysterixObservableCommand as completed if:
 - it hasn't timed out
 - it has a nonzero event count
 - it didn't fail to execute

There is also an added check at the end of markSuccess to fix the inconsistency of Status.CLOSED and circuitOpened. As described in https://github.com/Netflix/Hystrix/pull/1640, it is possible that: "After thread A change status from HALF_OPEN to CLOSED in markSuccess, thread B changed status from CLOSED to OPEN before thread A had a chance to reset the HealthCountsStream"