Fix HealthCounts creating new instance constantly
On every invocation of HealthCounts.plus() a new instance of HealthCounts
was being created leading to a lot of memory usage and gc impact. Because
HealthCounts.plus() is used as an unary aggregator in HealthCountsStream it
is not necessary for the aggregator to return a new instance each time and
it can instead simply return itself. Because the method mutates the state
of the object it must be synchronized to prevent racy updates to the fields
which could lead to incorrect computations of errorPercentage. This may
lead to a slight performance impact but is unlikely to have a major impact.

Additionally, empty() no longer returns a static reference but a new
instance each time.